Reference boolean settings by enum instead of string literals
Previously, settings were being accessed by name (using a string
literal), which means that every time a setting was accessed, its name
needed to be looked up in a string map. This is an unnecessary
waste of cycles, and becomes prohibitive to introducing preferences
that are to be accessed many times per frame.

Introduce a new enum in the Preferences class which enumerates all
boolean preferences (including whether a help message has been seen),
and update the rest of the code to use it. Aside from the performance
improvements, this has the benefit that a typo in a preference name
will result in a compilation error, instead of silently retrieving the
default value (false).

* Preferences.h:

  - Introduce Preferences::Preference enum (not scoped, thus in the
    Preferences class's namespace)

  - Update Has/Set methods to take preferences by enum instead of
    string

  - Add Name method to get a preference's name, as it appears in the
    same file

* Preferences.cpp:

  - Add/update implementations for the above methods

  - Replace the map<string,bool> with a bool[] for storing boolean
    settings

  - Move setting names into an array, so that they are queryable by
    e.g. Name()

* PreferencesPanel.cpp, PreferencesPanel.h:

  - Use integer indexes (an extension of the Preference enum) to
    identify clickable zones, instead of strings

  - Update to use the new Preferences API

  - DRY and fetch setting names from Preferences::Name() instead of
    duplicating them in the SETTINGS array

* Help.cpp, Help.h:

  - Introduce methods needed by the Preferences and PreferencesPanel
    classes to query and update the state of help texts.

* AI.cpp, DrawList.cpp, Engine.cpp, MainPanel.cpp, MapPanel.cpp,
  PlayerInfo.cpp, Sprite.cpp, main.cpp:

  - Use enum members to set and query settings instead of string
    literals